requirements.txt
cffi==1.2.1
Django==1.8.2
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
static==0.4
psycopg2==2.5.2
wsgiref==0.1.2
djangorestframework==3.1.3
django-push-notifications==1.3.1
django-storages==1.1.8
Pillow==2.9.0
boto==2.38.0
six==1.9.0
stripe==1.27.0
requests[security]==2.7.0
twilio==4.6.0

